modulized the code
