+++ src/data_models/exceptations/base.py
@@ -1,0 +1,4 @@
+class FaultTreeTextException(Exception):
+    def __init__(self, line_number, message):
+        self.line_number = line_number
+        self.message = message
