+++ src/data_models/nan.py
@@ -1,0 +1,9 @@
+class Nan:
+    """
+    Static class representing NaN (not a number).
+    """
+
+    def __new__(cls):
+        raise TypeError("`Nan` cannot be instantiated")
+
+    STRING = "nan"
